jit-elim: build categorized_sample_indices as int64, drop last te_Cast
The sampler does `sample_indices.long()` (sampler.py:794) on every decode
step. categorized_sample_indices was built with dtype=torch.int (int32),
so `.long()` dispatched a per-token int32->int64 cast which lowered to a
te_Cast TBE kernel on the NPU -- the last remaining JIT-compile trigger
in the Gemma4 path.

Build the index tensors directly as int64 (matching selected_token_indices
right above) so `.long()` is a no-op and no cast is dispatched. The CPU
tensor is created at the target dtype, then async_tensor_h2d does a
device-only H2D copy (no cast).

Cold-cache A/B (empty kernel_cache, 8-token greedy decode):
  int32 (before): 1 te_Cast compiled, total .o = 1
  int64 (after):  0 te_Cast,           total .o = 0
Identical output text. The full Gemma4 inference path now compiles zero
TBE kernels cold.

+++ vllm/model_executor/sampling_metadata.py
@@ -166,9 +166,11 @@ class SamplingMetadata:
             pin_memory=pin_memory,
         )
         categorized_sample_indices = {
+            # Build directly as int64 so sampler.py's `sample_indices.long()`
+            # is a no-op (avoids a per-token int32->int64 te_Cast on the NPU).
             t: async_tensor_h2d(
                 seq_ids,
-                dtype=torch.int,
+                dtype=torch.long,
                 target_device=device,
                 pin_memory=pin_memory,
             )
